fix bug:npu cannot calc with float64, cast to float32

+++ xt/model/muzero/muzero_model_ms.py
@@ -173,10 +173,10 @@ class MuzeroModelMS(XTModel_MS):
             self.reward_max)
         obs = Tensor.from_numpy(state[0])
         action = Tensor.from_numpy(state[1])
-        loss_weights = Tensor.from_numpy(state[2])
-        target_value = Tensor.from_numpy(target_value)
-        target_reward = Tensor.from_numpy(target_reward)
-        target_policy = Tensor.from_numpy(label[2])
+        loss_weights = Tensor.from_numpy(state[2]).astype(ms.float32)
+        target_value = Tensor.from_numpy(target_value).astype(ms.float32)
+        target_reward = Tensor.from_numpy(target_reward).astype(ms.float32)
+        target_policy = Tensor.from_numpy(label[2]).astype(ms.float32)
         loss = self.train_net(
             obs,
             action,
